Move volume in progress restore code into separate file
